+++ changelog.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added HasSingle as Linq extensions
 - Added TestBaseBuilder to permit build object for unit tests
 - Added LoggerMock
+- Added Health API which contains an extensions to use a custom response writer
 
 ## 0.4.0 &#8212; 2021-05-16